requirements.txt
streamlit
pandas
duckdb
langchain
langchain-openai
langchain-community
langgraph
python-dotenv
tabulate